+++ packages/connect/src/clients/createAppClient.ts
@@ -1,11 +1,16 @@
-import { connect, ConnectArgs, ConnectResponse } from "../actions/connect";
-import { status, StatusArgs, StatusResponse } from "../actions/status";
+import { connect, ConnectArgs, ConnectResponse } from "../actions/app/connect";
+import { status, StatusArgs, StatusResponse } from "../actions/app/status";
+import {
+  verifySignInMessage,
+  VerifySignInMessageArgs,
+  VerifySignInMessageResponse,
+} from "../actions/app/verifySignInMessage";
 import { Client, ClientConfig, createClient } from "./createClient";
-import { AsyncHttpResponse } from "./transports/http";
 
 export interface AppClient extends Client {
-  connect: (args: ConnectArgs) => AsyncHttpResponse<ConnectResponse>;
-  status: (args: StatusArgs) => AsyncHttpResponse<StatusResponse>;
+  connect: (args: ConnectArgs) => ConnectResponse;
+  status: (args: StatusArgs) => StatusResponse;
+  verifySignInMessage: (args: VerifySignInMessageArgs) => VerifySignInMessageResponse;
 }
 
 export const createAppClient = (config: ClientConfig): AppClient => {
@@ -14,5 +19,6 @@ export const createAppClient = (config: ClientConfig): AppClient => {
     ...client,
     connect: (args: ConnectArgs) => connect(client, args),
     status: (args: StatusArgs) => status(client, args),
+    verifySignInMessage: (args: VerifySignInMessageArgs) => verifySignInMessage(client, args),
   };
 };
