+++ packages/connect/src/clients/index.ts
@@ -4,10 +4,10 @@ export { createAuthClient } from "./createAuthClient";
 export type { AppClient } from "./createAppClient";
 export type { AuthClient } from "./createAuthClient";
 export type { ClientConfig } from "./createClient";
-export type { ConnectArgs, ConnectResponse } from "../actions/connect";
-export type { StatusArgs, StatusResponse } from "../actions/status";
+export type { ConnectArgs, ConnectResponse } from "../actions/app/connect";
+export type { StatusArgs, StatusResponse } from "../actions/app/status";
 export type {
   AuthenticateArgs,
   AuthenticateResponse,
-} from "../actions/authenticate";
+} from "../actions/auth/authenticate";
 export type { AsyncHttpResponse, HttpResponse } from "./transports/http";
