+++ packages/connect/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./errors";
-export * from "./messages";
 export * from "./clients";
