+++ packages/connect/src/messages/constants.ts
@@ -1,0 +1,2 @@
+export const STATEMENT = "Farcaster Connect";
+export const CHAIN_ID = 10;
