--- packages/connect/src/messages/index.ts
@@ -1,1 +1,0 @@
-export * as messages from "./messages";
